fixed Pathologist registration #100: Add the username and a logout button to the top-right area for all screens (not limited registration workflow).
On the "Professional" page of registration, please rewrite the questions
a. "Board certifications and specialties or equivalent for your country (or enter NA)"
b. "Board certification or residency organization and country (or enter NA)"

Please move the questions mentioned in 2. to the "Certifications" page.

Please force the current questions on the "Certifications" page to be integer.

I'm considering whether we should delete the screen name question on the "Certifications" page.

Finally, we'd like an email sent from caMicrscope to the pathologist acknowledging registration.